+++ daemon/libnetwork/types/types.go
@@ -12,8 +12,7 @@ import (
 	"github.com/moby/moby/v2/errdefs"
 )
 
-// constants for the IP address type
-// Deprecated: use the consts defined in github.com/docker/docker/libnetwork/resolvconf
+// constants for the IP address type.
 const (
 	IP = iota // IPv4 and IPv6
 	IPv4
